tests: Move threading helpers to dedicated file
