+++ tests/framework/layer_validation_tests.cpp
@@ -150,48 +150,6 @@ VKAPI_ATTR VkBool32 VKAPI_CALL DebugUtilsCallback(VkDebugUtilsMessageSeverityFla
     return VK_FALSE;
 }
 
-#if GTEST_IS_THREADSAFE
-void AddToCommandBuffer(ThreadTestData *data) {
-    for (int i = 0; i < 80000; i++) {
-        vk::CmdSetEvent(data->commandBuffer, data->event, VK_PIPELINE_STAGE_ALL_COMMANDS_BIT);
-        if (*data->bailout) {
-            break;
-        }
-    }
-}
-
-void UpdateDescriptor(ThreadTestData *data) {
-    VkDescriptorBufferInfo buffer_info = {};
-    buffer_info.buffer = data->buffer;
-    buffer_info.offset = 0;
-    buffer_info.range = 1;
-
-    VkWriteDescriptorSet descriptor_write = vku::InitStructHelper();
-    descriptor_write.dstSet = data->descriptorSet;
-    descriptor_write.dstBinding = data->binding;
-    descriptor_write.descriptorCount = 1;
-    descriptor_write.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-    descriptor_write.pBufferInfo = &buffer_info;
-
-    for (int i = 0; i < 80000; i++) {
-        vk::UpdateDescriptorSets(data->device, 1, &descriptor_write, 0, NULL);
-        if (*data->bailout) {
-            break;
-        }
-    }
-}
-
-#endif  // GTEST_IS_THREADSAFE
-
-void ReleaseNullFence(ThreadTestData *data) {
-    for (int i = 0; i < 40000; i++) {
-        vk::DestroyFence(data->device, VK_NULL_HANDLE, NULL);
-        if (*data->bailout) {
-            break;
-        }
-    }
-}
-
 void TestRenderPassCreate(ErrorMonitor *error_monitor, const vkt::Device &device, const VkRenderPassCreateInfo &create_info,
                           bool rp2_supported, const char *rp1_vuid, const char *rp2_vuid) {
     if (rp1_vuid) {
