+++ tests/unit/command_positive.cpp
@@ -14,7 +14,7 @@
 #include "../framework/layer_validation_tests.h"
 #include "../framework/pipeline_helper.h"
 #include "../framework/render_pass_helper.h"
-#include "../framework/thread_timeout_helper.h"
+#include "../framework/thread_helper.h"
 #include "generated/vk_extension_helper.h"
 
 TEST_F(PositiveCommand, SecondaryCommandBufferBarrier) {
