+++ tests/unit/sync_val_positive.cpp
@@ -14,7 +14,7 @@
 #include "../framework/layer_validation_tests.h"
 #include "../framework/pipeline_helper.h"
 #include "../framework/descriptor_helper.h"
-#include "../framework/thread_timeout_helper.h"
+#include "../framework/thread_helper.h"
 
 class PositiveSyncVal : public VkSyncValTest {};
